refactor SimpleVault.tsx to be pure sdk for contract.
